--- core/src/com/bisoft12/cenforpg/patterns/Behavioral/Estado/abstracto/Estado.java
@@ -1,7 +1,0 @@
-package com.bisoft12.cenforpg.patterns.Behavioral.Estado.abstracto;
-import com.bisoft12.cenforpg.patterns.Behavioral.Estado.objeto.Casa;
-
-public abstract class Estado {
-    public abstract String activarEstado();
-
-}//
