--- core/src/com/bisoft12/cenforpg/patterns/Comportamiento/Observador/Interfaces/Observador.java
@@ -1,5 +1,0 @@
-package com.bisoft12.cenforpg.patterns.Comportamiento.Observador.Interfaces;
-
-public interface Observador {
-    void irBatalla();
-}
