+++ core/src/com/bisoft12/cenforpg/patterns/Comportamiento/Patron_Estado/abstracto/Estado.java
@@ -1,0 +1,6 @@
+package com.bisoft12.cenforpg.patterns.Comportamiento.Patron_Estado.abstracto;
+
+public abstract class Estado {
+    public abstract String activarEstado();
+
+}//
