+++ core/src/com/bisoft12/cenforpg/patterns/Comportamiento/Patron_Observador/Interfaces/Observador.java
@@ -1,0 +1,5 @@
+package com.bisoft12.cenforpg.patterns.Comportamiento.Patron_Observador.Interfaces;
+
+public interface Observador {
+    void irBatalla();
+}
